Implement shard compatibility between H2 and MariaDB, add separate scheme for the H2, implement tests, collapse migration schemes, fix genesis account balance retrieval for shards with height < 1440, fix transaction_shard_index.partial_hash export (was 32 bytes, fixed to correct 24 bytes), add dbURL full support with sharding capabilities workaround, add docs for the apl-blockchain.properties

+++ apl-core/src/main/java/com/apollocurrency/aplwallet/apl/core/app/observer/CrowdFundingObserver.java
@@ -28,7 +28,6 @@ import lombok.extern.slf4j.Slf4j;
 import javax.enterprise.event.Observes;
 import javax.inject.Inject;
 import javax.inject.Singleton;
-import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -143,7 +142,7 @@ public class CrowdFundingObserver {
         FullTextOperationData operationData = new FullTextOperationData(
             DEFAULT_SCHEMA, currencyTable.getTableName(), Thread.currentThread().getName());
         operationData.setOperationType(operationType);
-        operationData.setDbIdValue(BigInteger.valueOf(currency.getDbId()));
+        operationData.setDbIdValue(currency.getDbId());
         operationData.addColumnData(currency.getName()).addColumnData(currency.getDescription());
         // send data into Lucene index component
         log.trace("Put lucene index update data = {}", operationData);
